requirements.txt
discord.py>=2.3.2
asyncpg
python-dotenv
